chore: refactor directory tree to put tlc specific code in a common tree
