+++ apps/automation-v2/src/main.ts
@@ -1,18 +1,11 @@
 import { createTaskRunner } from '@mountainpath9/overlord-core';
 
 import { batchLiquidate } from '@/tlc/batch-liquidate';
-
-import { CONFIG as CONFIG_TESTNETS } from '@/config/testnets';
-import { CONFIG as CONFIG_PRODNETS } from '@/config/prodnets';
-
+import { TLC_BATCH_LIQUIDATE_CONFIG } from '@/tlc/config';
 import { discordNotifyTaskException } from '@/common/discord';
 
 async function main() {
   const runner = createTaskRunner();
-
-  const mode = await runner.config.requireString('temple_tlc_networks');
-  const isProdnet = mode === 'prodnets';
-  const config = isProdnet ? CONFIG_PRODNETS : CONFIG_TESTNETS;
 
   runner.setVersion(process.env.VERSION || 'unknown');
   runner.setTaskExceptionHandler(discordNotifyTaskException);
@@ -20,7 +13,7 @@ async function main() {
   runner.addPeriodicTask({
     id: 'tlc-batch-liquidate',
     cronSchedule: '*/10 * * * *',
-    action: async (ctx) => await batchLiquidate(ctx, config.tlcBatchLiquidate),
+    action: async (ctx) => await batchLiquidate(ctx, TLC_BATCH_LIQUIDATE_CONFIG),
   });
 
   runner.main();
