+++ apps/automation-v2/src/tlc/config.ts
@@ -1,8 +1,19 @@
 import { parseEther } from 'ethers';
-import { MAINNET } from '@/chains';
-import { TlcBatchLiquidateConfig } from '@/tlc/batch-liquidate';
+import { Chain, TlcBatchLiquidateConfig } from '@/tlc/batch-liquidate';
 
-const TLC_BATCH_LIQUIDATE_CONFIG: TlcBatchLiquidateConfig = {
+export const MAINNET: Chain = {
+  id: 1,
+  name: 'Mainnet',
+  transactionUrl(txhash: string) {
+    return `https://etherscan.io/tx/${txhash}`;
+  },
+  addressUrl(address: string) {
+    return `https://etherscan.io/address/${address}`;
+  },
+};
+
+
+export const TLC_BATCH_LIQUIDATE_CONFIG: TlcBatchLiquidateConfig = {
   CHAIN: MAINNET,
   WALLET_NAME: 'temple_automation',
   TLC_ADDRESS: '0xcbc0A8d5C7352Fe3625614ea343019e6d6b89031',
@@ -12,8 +23,4 @@ const TLC_BATCH_LIQUIDATE_CONFIG: TlcBatchLiquidateConfig = {
   SUBGRAPH_URL:
     'https://subgraph.satsuma-prod.com/a912521dd162/templedao/tlc-liquidations-mainnet/api',
   SUBGRAPH_RETRY_LIMIT: 3,
-};
-
-export const CONFIG = {
-  tlcBatchLiquidate: TLC_BATCH_LIQUIDATE_CONFIG,
 };
